requirements.txt
pytorch-ignite==0.5.1
tensorboardX==2.6.2
torch==2.4.0
torch_geometric==2.5.3